increasing sampling rate

+++ src/gnss_task.h
@@ -9,7 +9,7 @@ typedef struct {
     float alt_m;
     uint8_t fix_type; // 0 = no fix, 1 = 2D fix, 2 = 3D fix
     uint8_t num_sats;
-    uint32_t timestamp;
+    int64_t timestamp;
 } gps_data_t;
 
 QueueHandle_t get_gps_queue();
